# Fix range deletion covering key in same SST file